constructor show candidates

+++ src/test/java/com/zergatul/scripting/tests/compiler/MethodOverloadTests.java
@@ -58,8 +58,8 @@ public class MethodOverloadTests extends ComparatorTest {
 
         String candidates = """
                 Candidates:
-                float m1(int x, int y, int z, string s)
-                float m1(float x, float y, float z, string s)""";
+                float m1(float x, float y, float z, string s)
+                float m1(int x, int y, int z, string s)""";
 
         comparator.assertDiagnostics(
                 ApiRoot.class, code, "⟦⟧",
